コメントを消せるユーザーの制限をCommentControllerのdestroyアクションメソッドに追加

+++ backend/app/Http/Controllers/CommentController.php
@@ -2,8 +2,10 @@
 
 namespace App\Http\Controllers;
 use App\Models\Comment;
+use App\Models\Plan;
 use Illuminate\Http\Request;
 use App\Http\Requests\CommentRequest;
+use Illuminate\Support\Facades\Auth;
 
 class CommentController extends Controller
 {
@@ -20,7 +22,12 @@ class CommentController extends Controller
 
     public function destroy(Comment $comment)
     {
-        $comment->delete();
+        $plan_id = $comment->plan_id;
+        $plan = Plan::where('id', $plan_id)->first();
+
+        if (Auth::id() === $comment->user_id || Auth::id() === $plan->user_id) {
+            $comment->delete();
+        }
 
         session()->flash('msg_success', 'コメントを削除しました');
         return redirect()->route('plans.show', ['plan' => $comment->plan_id]);
